second commit

+++ src/BonusMilesService.java
@@ -1,6 +1,6 @@
 public class BonusMilesService {
 
     public int calculate(int price) {
-        return (int) (price/20); // должно получиться 500
+        return price / 20; // должно получиться 500
     }
 }
